[UPDATE] example to dev
