+++ aon_a2a/configs.py
@@ -1,0 +1,2 @@
+from dotenv import dotenv_values
+config = dotenv_values("aon_a2a/.env")
