+++ aon_a2a/iac_agent/agent_executor.py
@@ -1,6 +1,6 @@
 from typing import override
 
-from examples.cloud_agent.agent import CloudAgent
+from aon_a2a.iac_agent.agent import IaCAgent
 
 from a2a.types import (
     TaskArtifactUpdateEvent,
@@ -11,13 +11,14 @@ from a2a.server.events import EventQueue
 from a2a.utils import new_text_artifact
 
 
-class CloudAgentExecutor(AgentExecutor):
+class IaCAgentExecutor(AgentExecutor):
 
     def __init__(self):
-        self.agent = CloudAgent()
+        self.agent = IaCAgent()
 
     @override
-    async def execute(self,
+    async def execute(
+        self,
         context: RequestContext,
         event_queue: EventQueue
     ) -> None:
