--- examples/configs.py
@@ -1,2 +1,0 @@
-from dotenv import dotenv_values
-config = dotenv_values("examples/.env")
